add(codegen/python): Add operation visitor implementation, to generate a basic SDK.
Output types are not perfect, since they use the generic return types from queries / mutations, as defined by the schema. Ideally, we would use a tailored type based on the actual selection set like the typescript SDK does.
For now, this requires adding a `everythingIsOptional` flag to the type visitor, which generates the output types with all fields optional