add(codegen/python): Add a utility method for converting the selection set for an operation into a tree structure of field names.
This is the first step towards being able to convert a selection set to concrete python types. In this step, we take take selection set, which includes field selections and fragment spreads, and converts it to only field selections. We also deduplicate field selections, so that each field is selected at most once at each step in the tree (multiple fragments might select the same field multiple times).
The implementation intentionally does not support inline fragments (yet). We are ~80% of the way to perfect output types without supporting inline fragments, so I want to handle them seperately