add(codegen/python): Add a utility method for converting the selection set for an operation into a tree structure of GraphQL AST nodes.
This is the second step towards being able to convert a selection set to concrete python types. In this step, we use the utility method from the previous step to convert the selection set to a tree structure of selected fields. We then extract an AST node representing the return type of the operation from the schema, and use the tree of selected fields to filter the fields of the AST node. Concretely, if the output type of a query returns two fields, `application` and `signatureOrder`, but the operation only selects `signatureOrder`, `operationSelectionsToAstTree` will return an AST node with only `signatureOrder` in its `fields` array.