change(codegen/python): Refactor visitors to be able to use `visit` from `graphql-js` instead of `oldVisit` from `@graphql-codegen/plugin-helpers`.
GraphQL codegen visitor work exclusively when leaving AST nodes. At one point you could specify a visitor like this, if you only needed to work when leaving a node
```ts
 {
  leave: {
    OperationDefinition(node: OperationDefinitionNode) {}
    FragmentDefinition(node: FragmentDefinitionNode) {}
  }
}
```
That no longer works in `graphql-js`, but was re-implenented in `@graphql-codegen/plugin-helpers`. However, the types are not narrowed correctly with how it's implemented, so we had to use both a `@ts-expect-error` and a type assertion.
To avoid that, we now specify visitors like this
```ts
{
  OperationDefinition: {
    enter(node: OperationDefinitionNode) {},
    leave(node: OperationDefinitionNode) {},
  },
  FragmentDefinition: {
    enter(node: FragmentDefinitionNode) {},
    leave(node: FragmentDefinitionNode) {},
  },
}
```
Which is slightly more WET, but avoids disabling the type-check