change(codegen/python): Remove visitor functions for `NamedType`, `NonNullType` and `ListType`.
Generate everything that has to do with type names, list types and non-nulls in `FieldOrInputValueDefinition` instead. We need more control over how we generate fields.
This allows us to e.g. use different scalar types for input and output types (coming in a later commit).
By using this approach we also avoid the weirdness of always marking types as optional in `NamedType`, and then removing the modifier again if the type is wrapped in a `NonNullType` further up the tree.

This commit changes nothing about the output of the codegen - that comes later.