change(codegen/python): In `createAstTreeNodeFromSelection`, handle fragments in the selection object, and create separate types for each concrete type implementing an interface.
* Previously `collectSelectedFieldsFromOperation` would only return objects and scalars. Now it also returns fragments, and we need to consider `fragment.fields` when checking if a field has been selected on a type.
* Previously, we did not account for the fact that interfaces can have different subtypes, each with different field selections. Now, whenever we encounter a node which is an interface, we loop through all the types which implement this interface, and filter their fields according to the selection.

Take note, that the core loop `for (const field of astNode.fields)` remains unchanged. The way in which we filter fields on AST nodes according to the selection is still the same. The addition here is that we filter the fields from multiple possible types, when the `schemaNode` is an interface.