change(codegen/python): Use RootModel to parse result.
Most operations will return a concrete type, which we can call `model_validate` on. However, `querySignatureOrders` uses the `viewer` interface, and thus the return type is a union.
In order to support both concrete models and interfaces, we wrap in a pydantic `RootModel`, in order to be able to parse it.