change(codegen/python): Make all SDK methods async
Coming from node.js I've also learned "don't block the event loop!". This seems like a good idea in python as well https://realpython.com/async-io-python/#when-to-use-async-io.
It remains to be seen whether all SDK consumers can / want to use an async version of the SDK, but let's cross that bridge when we get there.