change(codegen/python): Create type guards for types that implement interfaces.
When a GraphQL query returns an interface type, we need some way to distinquish which concrete type an object has. Previously, we would use an `instanceof` check, but this leaks implementation details, since you have to check against a generated name such as `CreateSignatureOrder_CreateSignatureOrderOutput_SignatureOrder_Document_PdfDocument`.

Ideally consumers should not have to implement the generated classes.

So, we create type guards instead, which helps to narrow the type. See `test_operations.py` for an example.