change(packages/python): Return graphql blob as bytes, instead of strings.
The fact that blobs are base64 encoded is an implementation detail of our API. As an API consumer, I expect to get some bytes back that I can easily save to a file or send back to the client.
This also matches the fact that we allow passing in bytes when creating a signature order.