change(packages/python): Simplify the return type from `querySignatureOrders`.
In the GraphQL schema, we query signature orders on the viewer. The viewer can have many different types, but we know that in an SDK context, `viewer` will always return an application viewer.

Furthermore we also remove the notion of edges and nodes from the return type, and just return a list of signature orders. This does make it harder to do pagination, since we loose the `pageInfo` and `cursor`, but it is in line with what we do in the node.js SDK.