change(packages/python): Replace ty with pyright.
ty is fast, and nice to work with. However, its inference/type narrowing around generics is not strong enough yet. See https://github.com/astral-sh/ty/issues/1292 for details.
Instead of jumping through hoops to work around that limitation, this PR replaces ty with pyright.

pyright has a `--project` flag similar to ty. However, it doesn't work as well, so instead we are placing a package.json with a lint-staged section in both the python and the example packages. This ensures that we run pyright in the right context, with the right dependencies avaiable