change(codegen/python/types): Annotate deprecated fields
Adding `deprecated` to the pydantic field config logs a deprecation warning when the field is accessed. However, we also want to show a deprecation warning to type-checkers. For that, we need to add a property with the @deprecated annotiation.

So, for a field called `role`, we add a backing property called `roleDeprecated`, with alias `role`. This is the property that is parsed by pydantic.
We then add a property with a getter called `role`, and mark it as deprecated. The getter calls `model_dump`, and returns the `roleDeprecated´ field.