change(packages/nodejs): Add explicit return types to operations.
Without this, the return types is inline, so we are essentially duplicating the type, so it exists both in the application-viewer-types.ts file, and in the application-viewer.ts file. Now the hand-written part of the SDK uses the generated types directly.

Inlining the types also means we loose any comments, such as @deprecated on fields.