change(codegen/nodejs): Make it possible to mark fields as deprecated.
We would like to add the @deprecated jsdoc tag to fields that are marked as deprecated in our GraphQL schema. Unfortunately, that is not supported by the typescript codegen plugin. So we 'hack' it a bit, by applying a codemod transformation after the code has been generated. This does mean that the deprecations are not actually tied to which fields are marked as deprecated in the schema, but it should serve us for now.

The codemod matches fields based on name, but does not consider any context (i.e., which type the field appears in) - again, this should serve us for now.

This commit also marks the role field as deprecated